requirements.txt
torch>=2.0
torchvision
timm
pandas
numpy
scikit-learn
opencv-python
Pillow
streamlit
albumentations
albumentations-pytorch
matplotlib
scipy
joblib
